// WEB/src/app/modules/_auth/login/model/taikhoan-view-model.ts
export class TaiKhoanViewModel {
    ID_TaiKhoan: number = 0;
    TenDangNhap: string ='';
    MatKhau: string ='';
    XacNhanMatKhau: string ='';
    TenDayDu: string ='';
    IsKhachHang: boolean = false;
    GioiTinh: string ='NAM';
    IsLock: boolean = false;
    Email: string ='';
    SoDienThoai: string ='';
    ID_ChucDanh: number = 0;
    ID_TinhThanh: number = 0;
    ID_QuanHuyen: number = 0;
    ID_PhuongXa: number = 0;
}